disable reporting exceptions to exceptional when using laptop.

+++ server/code/index.coffee
@@ -648,9 +648,10 @@ process.on 'SIGTERM', ->
     process.exit 1
   , 30*1000
 
-process.on 'uncaughtException', (err) ->
-  console.warn err
-  Exceptional.handle err
-  setTimeout ->
-    process.kill process.pid, 'SIGTERM'
-  , 500
+if /staging|production/.test process.env.NODE_ENV
+  process.on 'uncaughtException', (err) ->
+    console.warn err
+    Exceptional.handle err
+    setTimeout ->
+      process.kill process.pid, 'SIGTERM'
+    , 500
